Time: 0 ms (100.00%) | Memory: 11.9 MB (29.14%) - LeetSync

+++ 63-unique-paths-ii/unique-paths-ii.cpp
@@ -23,7 +23,7 @@ public:
         }
     }
 
-    return (int)dp[0][0]; // if return type must be int, cast back
+    return dp[0][0]; // if return type must be int, cast back
 }
 
 };
